Process Copy and Code renditions as arrays
The previous approach involved simply concatenating all of the pieces of the content that either were or weren’t code. This means that, effectively, new strings were formed by that concatenation, and the library would find strings that didn’t actually exist in the source. The library now reviews each snippet individually instead of concatenating.